Try to make the warning a little bit more accurate in netCDF data store. When the CRS is declared in different ways, verify that they are consistent. Store enumeration values on 64 bits integers instead of 32 bits, because they are sometime bitmask with values that require 32 bits unsigned integers.

+++ endorsed/src/org.apache.sis.storage.netcdf/main/org/apache/sis/storage/netcdf/base/FeatureSet.java
@@ -900,12 +900,12 @@ makeGeom:   if (!isEmpty) {
                     } else {
                         value = p.read(extent, null);               // Force the type to `Vector`.
                     }
-                    final Map<Integer,String> enumeration = p.getEnumeration();
+                    final Map<Long,String> enumeration = p.getEnumeration();
                     if (enumeration != null && value instanceof Vector) {
                         final var data = (Vector) value;
                         final var meanings = new String[data.size()];
                         for (int j=0; j<meanings.length; j++) {
-                            String m = enumeration.get(data.intValue(j));
+                            String m = enumeration.get(data.longValue(j));
                             meanings[j] = (m != null) ? m : "";
                         }
                         value = Arrays.asList(meanings);
